Web design
Please contact me to view the app design as it was too large to upload to GitHub. All though the project did not pass the prototype phase, a lot was learnt about launching a start up, Adobe Xd, Adobe illustrator, the market research needed to launch and app, pitch decks, funding rounds, team work and design practices. Both the web and app can be vastly improved, this was just to test the waters and used as a learning experience to help guide us into what we want to work in which led to the discovery of tableau. The designs were created between a university friend and myself.